algorithms_python/recommendation_engine/update-count-matrix/requirements.txt
numpy==1.19.5
pandas==0.23.4
nltk==3.4
protobuf==3.17.3
scikit_learn==0.24.2
google-cloud-firestore==2.3.1
